move mix_with_colour to colour space namespace
